Fix fp16 gradient unscaling error during SFT training
Load model in float32 instead of float16. The fp16=True flag in
TrainingArguments handles mixed precision via autocast + GradScaler.
Loading the model directly in fp16 causes GradScaler to fail because
it cannot unscale already-fp16 gradients. Total VRAM ~9.4GB, still
fits T4 comfortably.

+++ mathphd_plus_plus/configs/base_config.py
@@ -11,7 +11,7 @@ from typing import Optional, List
 class ModelConfig:
     """Base model configuration."""
     model_name: str = "Qwen/Qwen2.5-0.5B"
-    torch_dtype: str = "float16"  # T4 doesn't support bf16
+    torch_dtype: str = "float32"  # Load in fp32; fp16 mixed precision handled by TrainingArguments
     use_flash_attention: bool = False  # T4 doesn't support flash-attn2
     gradient_checkpointing: bool = True
     max_seq_length: int = 2048
